Move test to test_main.py
