+++ main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/pyhton
 
-import requests, time, datetime, sys, os
+import requests, time, datetime, sys
 from dateutil.relativedelta import relativedelta
 
 def main(args):
@@ -72,19 +72,6 @@ def main(args):
 
         now_datetime = now_datetime + datetime.timedelta(days=-1)
     return found_count
-
-def test_main():
-    args = []
-    args.append('github_id='+os.environ['MY_GITHUB_ID'])
-    args.append('github_token='+os.environ['MY_GITHUB_TOKEN'])
-    args.append('search_topic=hacktoberfest-dummy-test')
-    args.append('search_month_range=0')
-    args.append('search_day_range=1')
-    args.append('search_start_date=2020-11-01')
-    args.append('search_location=Korea')
-    
-    found_count = main(args)
-    assert found_count == 1
 
 if __name__ == "__main__":
     main(sys.argv)
